Instead of deleting APRUploadService replace it with a dummy
Decompiling SystemUI has shown that it also contains some data collection, that
is automatically activated when APRUploadService is not installed. If it is
installed, it gets queried for a disable / enable flag. By replacing the app
with a dummy that only shares the name, we trigger the check for the settings
flag, which then fails and makes the collection check default to false.